cmd/plugins/transformer/pdf/requirements.txt
grpcio
grpcio-tools
requests
pypdf